Move `URLProtocolStub` to a separate file
